requirements.txt
## necessary
imagehash
opencv-python
distance
you-get

## for server 
flask
gitpython